% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.wave4.R
\name{d.wave}
\alias{d.wave}
\title{\code{d.wave} 2D lunge detector for individual dive, with validation}
\usage{
d.wave(data = df, dive = 96, presmooth.window = 4, pk.per = 16,
  per.rng = 20, min.pkdist = 40, a.pk = 0.05, pk.env = 0.85,
  return.wavelet = F, outliers = "flag", edges = "flag",
  plotting = T)
}
\arguments{
\item{data}{Data frame containing dive info (e.g. created using \code{read.depth}, \code{read.lunges},
\code{join.dstamp} and \code{number.dive}).}

\item{dive}{Sequence number of dive to be analysed.}

\item{presmooth.window}{Initial window of smooth for gaussian running mean filter}

\item{pk.per}{Peak wavelet period for weighted power analysis}

\item{per.rng}{Range of all periods to be considered in weighted power analysis}

\item{min.pkdist}{Minimum period between consecutive peaks}

\item{a.pk}{Minimum amplitude for peak to be considered significant}

\item{pk.env}{Proportion of a.pk, used to set the duration of what is considered a peak}

\item{return.wavelet}{Logical, whether wavelet object should be returned or only the detected peaks}

\item{outliers}{Logical, either flag outliers or delete them (flag by default, change to 'omit' to omit)}

\item{edges}{Logical, either flag edge peaks or delete them (flag by default, change to 'omit' to omit)}

\item{plotting}{Whether to plot the results or not}
}
\value{
Returns a list with objects \code{parameters}, containing all parameters used when running detector,
\code{wavelet} containing the wavelet output (if \code{return.wavelet} was set to TRUE) and \code{peaks} data frame containing the detected peaks
}
\description{
\code{d.wave} 2D lunge detector for individual dive, with validation
}
\examples{
d <- read.depth(whalename='mn16_Jan25b')
t <- read.lunges(whalename='mn16_Jan25b')
df <- join.dstamp(d, t)
df$dive <- number.dive(df$depth)
d103 <- d.wave(dive=103)
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{read.depth}} to read depth time series,
  \code{\link{join.dstamp}} to merge lunge events and dive data,
  \code{\link{number.dive}} to add dive numeric ID column,
  \code{\link{run.d.wave}} to run \code{d.wave} on entire data record from an animal.

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.depth}},
  \code{\link{read.lunges}}, \code{\link{run.d.wave}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
