% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.wave.TDR.R
\name{d.wave.TDR.original}
\alias{d.wave.TDR.original}
\title{\code{d.wave.TDR.original} 2D lunge detector for individual dive in a TDR record. Original backup version}
\usage{

  d.wave.TDR.original(tdr.file = "./Data/HVTag/Rfiles/Whale_2015Feb23.RData",
  tdr.folder = "./Data/HVTag/Rfiles")
}
\arguments{
\item{tdr.file}{R object containing pre-processed dive record.
If set to NA, this function can run the detector on data from all individuals
with data records stored in the directory specified in \code{tdr.folder}.}

\item{tdr.folder}{Folder in which TDR files are stored.}
}
\description{
\code{d.wave.TDR.original} 2D lunge detector for individual dive in a TDR record. Original backup version
}
\details{
Run the 2D lunged detector on TDR-type dive data, i.e. this version has no validation against pre-determined lunges
}
\examples{
tdr.list <- d.wave.TDR.original(tdr.file=NA)
}
\seealso{
\code{\link{run.d.wave}} for data type with validation against pre-determined lunges.

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.depth}},
  \code{\link{read.lunges}}, \code{\link{run.d.wave}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
