% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.lunges.R
\name{detect.lunge}
\alias{detect.lunge}
\title{\code{detect.lunge} Deprecated lunge detection function}
\usage{
detect.lunge(data = dfc, method = "resid", window = 4, square = F,
  q.pk = NA, a.pk = 0.15, d.cut = 0, pkdist = 20, plotting = F)
}
\description{
\code{detect.lunge} Deprecated lunge detection function
}
\author{
Martin Biuw
}
