% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.lunges.R
\name{dstamp.plot}
\alias{dstamp.plot}
\title{\code{dstamp.plot} Plot dive record with pre-determined lunges}
\usage{
dstamp.plot(df)
}
\arguments{
\item{df}{Data frame with data in dstamp format}
}
\value{
Produces simple plot of dives and lunges
}
\description{
\code{dstamp.plot} Plot dive record with pre-determined lunges
}
\examples{
dstamp.plot()
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{read.depth}} to read dive data,
  \code{\link{number.dive}} to add dive numeric ID column

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
