% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.wave.TDR.R
\name{filter.lunges}
\alias{filter.lunges}
\title{\code{filter.lunges} Filter out lunges during short & shallow dives}
\usage{
filter.lunges(data = tdr.df, duration = 60, depth = 10)
}
\arguments{
\item{data}{Output from running \code{d.wave.TDR}.}

\item{duration}{Maximum dive duration of dives considered for filtering out lunges}

\item{depth}{Maximum dive depth of dives considered for filtering out lunges}
}
\description{
\code{filter.lunges} Filter out lunges during short & shallow dives
}
\examples{
tdr.list <- d.wave.TDR(tdr.file=NA)
}
\seealso{
\code{\link{d.wave.TDR}} to run 2D detector on TDR data.

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
