% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.rene.R
\name{gpsWhale}
\alias{gpsWhale}
\title{\code{gpsWhale} Read Fastloc gps data from ESRI shapefile.}
\usage{
gpsWhale(filename = "./Data/Rene data/shp/Mn-2014-Dec-16.shp",
  crs = "+proj=utm +zone=33 +ellps=intl +units=m +no_defs")
}
\arguments{
\item{crs}{Coordinate Reference System (CRS) to be used (default is UTM zone 33, appropriate for Norwegian coast).}

\item{file}{File name (including relative path) of file to be read.}
}
\value{
Returns SpatialPointsDataFrame
}
\description{
\code{gpsWhale} Read Fastloc gps data from ESRI shapefile.
}
\seealso{
\code{\link{matLunge}} to read lunge data,
\code{\link{matDive}} to read dive data,
\code{ncWhale} to read data from standard netCDF tagtools file,
\code{whale2dstamp} to convert to dstamp format that can be used in 2D lunge detection with validation (e.g. \code{run.d.wave})

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.depth}},
  \code{\link{read.lunges}}, \code{\link{run.d.wave}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
