% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.all.rene.R
\name{import.all.rene}
\alias{import.all.rene}
\title{\code{import.all.rene} Import all data records from Dtag /Little Leonardo deployments}
\usage{
import.all.rene(directory = "./Data/Rene data/lunge_res",
  gps.dir = "./Data/Rene data/shp",
  gpsCRS = "+proj=utm +zone=33 +ellps=intl +units=m +no_defs")
}
\arguments{
\item{directory}{Directory where dive /accelerometer data are stored}

\item{gps.dir}{Directory where GPS data are stored}

\item{crs}{Coordinate Reference System (CRS) to be used (default is UTM zone 33, appropriate for Norwegian coast).}
}
\value{
Returns a list with components \code{raw}, containing the raw imported dive/accelerometer data,
\code{dive} containing specific dive data, \code{lunge} containing lunge data
}
\description{
\code{import.all.rene} Import all data records from Dtag /Little Leonardo deployments
}
\details{
NOTE! This function creates new objects in the working environment automatically,
with names corresponding to the whale deployment names. No need therefore to save the results
into a new object.
}
\examples{
import.all.rene()
## Then add new object \\code{ds} to each resulting list, in dstamp format ready for d.wave:
file.pos <- ls()[grep('mn1', ls())]
for(i in file.pos) eval(parse(text=paste(i, '$ds <- whale2dstamp(', i, ')', sep='')))
}
\seealso{
\code{\link{matLunge}} to read lunge data,
\code{\link{matDive}} to read dive data,
\code{ncWhale} to read data from standard netCDF tagtools file,
\code{gpsWhale} to read GPS position data,
\code{whale2dstamp} to convert to dstamp format that can be used in 2D lunge detection with validation (e.g. \code{run.d.wave})

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{d.wave}},
  \code{\link{dstamp.plot}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.depth}},
  \code{\link{read.lunges}}, \code{\link{run.d.wave}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
