% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.lunges.R
\name{join.dstamp}
\alias{join.dstamp}
\title{\code{join.dstamp} Merge depth and lunge data into data.frame}
\usage{
join.dstamp(d, t, trim = T)
}
\arguments{
\item{d}{Dive data, as created using \code{read.depth}}

\item{t}{Lunge data, as created using \code{read.lunge}}
}
\value{
Returns a data frame with variables \code{time},
\code{depth} and \code{lunge}
}
\description{
\code{join.dstamp} Merge depth and lunge data into data.frame
}
\examples{
t <- join.dstamp(d,t)
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{read.depth}} to read dive data,
  \code{\link{number.dive}} to add dive numeric ID column,

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
