% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.lunges.R
\name{match.lunges}
\alias{match.lunges}
\title{\code{match.lunges} Lunge matching function}
\usage{
match.lunges(data = df, window = 3)
}
\arguments{
\item{data}{Data frame with pre-determined lunges and detected peaks to be matched}

\item{window}{Period before and after detected peak within which pre-determined peaks are considered matched}
}
\value{
list with 2 objects: \code{matched}, logical vector indicating which peaks were matched (or not) to a pre-determined lunge,
\code{confusion}, confusion matrix summarizing classification accuracy
}
\description{
\code{match.lunges} Lunge matching function
}
\details{
NOTE! THis function is called by \code{run.d.wave}, so usually does not have to be run
}
\examples{
matched <- match.lunges(df)
}
\seealso{
\code{\link{run.d.wave}} to run 2D detector on entire data record,

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
