% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.rene.R
\name{matDive}
\alias{matDive}
\title{\code{matDive} Read dive data from acceleration data logger stored in Matlab format}
\usage{

  matDive(file = "./Data/Rene data/lunge_res/03_Herring_mn13_340a_herrdive.mat")
}
\arguments{
\item{file}{File name (including relative path) of file to be read.}
}
\value{
Returns list with components representing dives
}
\description{
\code{matDive} Read dive data from acceleration data logger stored in Matlab format
}
\seealso{
\code{\link{matLunge}} to read lunge data,
\code{ncWhale} to read data from standard netCDF tagtools file,
\code{gpsWhale} to read GPS position data,
\code{whale2dstamp} to convert to dstamp format that can be used in 2D lunge detection with validation (e.g. \code{run.d.wave})

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{d.wave}},
  \code{\link{dstamp.plot}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.depth}},
  \code{\link{read.lunges}}, \code{\link{run.d.wave}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
