% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkDygraph.R
\name{mkDygraph}
\alias{mkDygraph}
\title{\code{mkDygraph} Make interactive dygraph of data with (optional) detected lunges.}
\usage{
mkDygraph(l.data = tst)
}
\arguments{
\item{l.data}{List of data obtained from \code{run.d.wave}.}
}
\value{
Creates an interactive plot where pre-determined lunges (if present), detected peaks (if present) and matched lunges & peaks.
}
\description{
\code{mkDygraph} Make interactive dygraph of data with (optional) detected lunges.
}
\details{
Pre-determined lunges (if present) are indicated by red solid points.
Peaks detected using the 2D detector (if present) are represented by green solid points.
Matched lunges/peaks are enclosed in purple diamonds.
}
\examples{
d <- read.depth(whalename='mn16_Jan25b')
t <- read.lunges(whalename='mn16_Jan25b')
df <- join.dstamp(d, t)
df$dive <- number.dive(df$depth)
df <- run.d.wave(df)
mkDygraph(df)
}
\seealso{
\code{\link{run.d.wave}} to run \code{d.wave} on entire data record from an animal.

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
