% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.d.wave.R
\name{optim.d.wave}
\alias{optim.d.wave}
\title{\code{optim.d.wave} Obtain ptimize}
\usage{
optim.d.wave(par, theData = df)
}
\arguments{
\item{par}{Vector of starting values for parameters to be estimated}

\item{theData}{Data frame containing dive info (e.g. created using \code{read.depth}, \code{read.lunges},
\code{join.dstamp} and \code{number.dive}).}
}
\value{
Returns the objective function for optimal parameter combinations
}
\description{
\code{optim.d.wave} Obtain ptimize
}
\details{
NOT YET FULLY IMPLEMENTED. EXTREMELY TIME CONSUMING!
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{read.depth}} to read depth time series,
  \code{\link{join.dstamp}} to merge lunge events and dive data,
  \code{\link{number.dive}} to add dive numeric ID column,
  \code{\link{d.wave}} to run lunge detector omn single dives.
  \code{\link{run.d.wave}} to run \code{d.wave} on entire data record from an animal.

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{d.wave}},
  \code{\link{dstamp.plot}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{read.depth}},
  \code{\link{read.lunges}}, \code{\link{run.d.wave}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
