% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.lunges.R
\name{read.depth}
\alias{read.depth}
\title{\code{read.depth} Read depth data}
\usage{
read.depth(dirname = "./Data/Takashi data/", whalename = "mn16_Jan26",
  toff = 1)
}
\arguments{
\item{whalename}{Name of the whale for which data are to be retrieved}
}
\value{
Returns a list with objects \code{start} (start time),
\code{freq} (sampling frequency) and \code{d} (vector of depth data)
}
\description{
\code{read.depth} Read depth data
}
\examples{
t <- read.depth()
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{join.dstamp}} to merge lunge events and dive data,
  \code{\link{number.dive}} to add dive numeric ID column,

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
