% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.lunges.R
\name{read.lunges}
\alias{read.lunges}
\title{\code{read.lunges} Read positions of lunges}
\usage{

  read.lunges(filename = "./Data/Takashi data/Time stamp of Lunge for Norway.xlsx",
  whalename = "mn16_Jan19a")
}
\arguments{
\item{filename}{Name of the Excel file containing lunge data}

\item{whalename}{Name of the sheet from which to read data}
}
\value{
Returns a list with objects \code{start} (start time),
\code{freq} (sampling frequency) and \code{stamps} (index of stamp occurrences)
}
\description{
\code{read.lunges} Read positions of lunges
}
\details{
Note! R (or RStudio) must be run in 32-bit mode for this function to work.
}
\examples{
t <- read.lunges()
}
\seealso{
\code{\link{read.depth}} to read depth data,
  \code{\link{join.dstamp}} to merge lunge events and dive data,
  \code{\link{number.dive}} to add dive numeric ID column,

Other Lunge detector functions: \code{\link{d.wave.TDR}},
  \code{\link{d.wave}}, \code{\link{dstamp.plot}},
  \code{\link{filter.lunges}}, \code{\link{gpsWhale}},
  \code{\link{import.all.rene}}, \code{\link{join.dstamp}},
  \code{\link{matDive}}, \code{\link{matLunge}},
  \code{\link{match.lunges}}, \code{\link{mkDygraph}},
  \code{\link{ncWhale}}, \code{\link{number.dive}},
  \code{\link{optim.d.wave}}, \code{\link{read.depth}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}},
  \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
