% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.d.wave.R
\name{run.d.wave}
\alias{run.d.wave}
\title{\code{d.wave} 2D lunge detector for entire record, with validation}
\usage{
run.d.wave(name = NA, data = NULL, show.info = F, plotting = T,
  match.window = 15, smooth.window = 4, pper = 16, rper = 20,
  min.pdist = 40, min.ppow = 0.05, pk.e = 0.85)
}
\arguments{
\item{name}{Name of animal to analyse, if data are to be read from external files (default is NA)}

\item{data}{Data frame containing dive info (e.g. created using \code{read.depth}, \code{read.lunges},
\code{join.dstamp} and \code{number.dive}).}

\item{show.info}{Logical, should detaild information be shown while running function}

\item{plotting}{Whether to plot the results or not}

\item{match.window}{Size of time window within which peaks and lunges should be considered matched}

\item{smooth.window}{Initial window of smooth for gaussian running mean filter (equivalent to \code{presmooth.window} in \code{d.wave})}

\item{pper}{Peak wavelet period for weighted power analysis (equivalent to \code{pk.per} in \code{d.wave})}

\item{rper}{Range of all periods to be considered in weighted power analysis (equivalent to \code{per.rng} in \code{d.wave})}

\item{min.pdist}{Minimum period between consecutive peaks (equivalent to \code{min.pkdist} in \code{d.wave})}

\item{min.ppow}{Minimum amplitude for peak to be considered significant (equivalent to \code{a.pk} in \code{d.wave})}

\item{pk.e}{Proportion of a.pk, used to set the duration of what is considered a peak (equivalent to \code{pk.env} in \code{d.wave})}
}
\value{
Returns a list with objects \code{parameters}, containing all parameters used when running detector,
\code{df} data frame containing original data plus peak detections and matching information, \code{detects} list containing wavelet ojects for each dive,
\code{matched} list containing 1) a logical vector indicating the occurrence of matched peaks/lunges, and 2) confusion matrix for matched and unmatched peaks/lunges
}
\description{
\code{d.wave} 2D lunge detector for entire record, with validation
}
\examples{
d <- read.depth(whalename='mn16_Jan25b')
t <- read.lunges(whalename='mn16_Jan25b')
df <- join.dstamp(d, t)
df$dive <- number.dive(df$depth)
df <- run.d.wave(data=df)
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{read.depth}} to read depth time series,
  \code{\link{join.dstamp}} to merge lunge events and dive data,
  \code{\link{number.dive}} to add dive numeric ID column,
  \code{\link{d.wave}} to run lunge detector omn single dives.

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{d.wave}},
  \code{\link{dstamp.plot}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{whale2dstamp}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
