% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.all.rene.R
\name{whale2dstamp}
\alias{whale2dstamp}
\title{\code{whale2dstamp} Convert data from object created by \code{import.all.rene} to data frame compatible with \code{d.wave} and \code{run.d.wave}.}
\usage{
whale2dstamp(lst = mn13_340a)
}
\arguments{
\item{lst}{Object obtained from \code{ncWhale}}
}
\value{
Returns data frame that can be used in 2D lunge detector with validation
}
\description{
\code{whale2dstamp} Convert data from object created by \code{import.all.rene} to data frame compatible with \code{d.wave} and \code{run.d.wave}.
}
\seealso{
\code{import.all.rene},
\code{\link{matLunge}} to read lunge data,
\code{\link{matDive}} to read dive data,
\code{ncWhale} to read data from standard netCDF tagtools file,
\code{gpsWhale} To read Fastloc GPS data

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{d.wave}},
  \code{\link{dstamp.plot}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{wt.image.plot}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
