% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.wave4.R
\name{wt.image.plot}
\alias{wt.image.plot}
\title{\code{wt.image.plot} Plot wavelet image for single dive}
\usage{
wt.image.plot(dw.obj)
}
\arguments{
\item{dw.obj}{Object (output from \code{d.wave}) for which wavelet image should be plotted}
}
\value{
Returns a colour image showing the power across time (horizontal) and wavelet periods (vertical)
}
\description{
\code{wt.image.plot} Plot wavelet image for single dive
}
\examples{
d <- read.depth(whalename='mn16_Jan25b')
t <- read.lunges(whalename='mn16_Jan25b')
df <- join.dstamp(d, t)
df$dive <- number.dive(df$depth)
d103 <- d.wave(dive=103)
plot.wt.image(d103)
}
\seealso{
\code{\link{read.lunges}} to read lunge event timestamps,
  \code{\link{read.depth}} to read depth time series,
  \code{\link{join.dstamp}} to merge lunge events and dive data,
  \code{\link{number.dive}} to add dive numeric ID column,
  \code{\link{d.wave}} 2D lunge detector for individual dive

Other Lunge detector functions: \code{\link{d.wave.TDR.original}},
  \code{\link{d.wave.TDR}}, \code{\link{d.wave}},
  \code{\link{dstamp.plot}}, \code{\link{filter.lunges}},
  \code{\link{gpsWhale}}, \code{\link{import.all.rene}},
  \code{\link{join.dstamp}}, \code{\link{matDive}},
  \code{\link{matLunge}}, \code{\link{match.lunges}},
  \code{\link{mkDygraph}}, \code{\link{ncWhale}},
  \code{\link{number.dive}}, \code{\link{optim.d.wave}},
  \code{\link{read.depth}}, \code{\link{read.lunges}},
  \code{\link{run.d.wave}}, \code{\link{whale2dstamp}}
}
\author{
Martin Biuw
}
\concept{Lunge detector functions}
